feat: refactor config, support plasmid features
This commit adds the ability to specify features (specifically,
resistance genes and origins of replication) for plasmids.  While making
this change, I realized that the options read from the config file
should be stored in the database object, and not in a global variable.
I fixed this as well, and while I was at it I made the config file
system more flexibile in terms of file/directory layout and reagent
naming conventions (both things I'd been meaning to do for a while).

Fixes #10
Fixes #12